Use another thread for webhook
