+++ pogom/utils.py
@@ -86,23 +86,3 @@ def get_encryption_lib_path():
         raise Exception(err)
 
     return lib_path
-
-# Add webhook for PokeAlarm -- Ugly Copy Pasta
-def send_to_webhook(message_type, message):
-    args = get_args()
-
-    data = {
-        'type': message_type,
-        'message': message
-    }
-
-    if args.webhooks:
-        webhooks = args.webhooks
-
-        for w in webhooks:
-            try:
-                requests.post(w, json=data, timeout=(None, 5))
-            except requests.exceptions.ReadTimeout:
-                log.error('Response timeout on webhook endpoint %s', w)
-            except requests.exceptions.RequestException as e:
-                log.error(e)
